+++ src/main/java/net/dries007/tfc/client/render/animal/RenderBearTFC.java
@@ -34,7 +34,7 @@ public class RenderBearTFC extends RenderLiving<EntityBearTFC>
     @Override
     public void doRender(@Nonnull EntityBearTFC bear, double par2, double par4, double par6, float par8, float par9)
     {
-        this.shadowSize = 0.35f + (bear.getPercentToAdulthood() * 0.35f);
+        this.shadowSize = (float) (0.35f + (bear.getPercentToAdulthood() * 0.35f));
         super.doRender(bear, par2, par4, par6, par8, par9);
     }
